src/chatbotservice/requirements.txt
## from src/recomendationservice/requirements.in

cachetools==3.1.1         # via google-auth
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
google-api-core[grpc]==1.22.0  # via -r requirements.in, google-api-python-client, google-cloud-core, google-cloud-trace, opencensus
google-api-python-client==1.10.0  # via google-cloud-profiler, google-python-cloud-debugger
google-auth-httplib2==0.0.4  # via google-api-python-client, google-cloud-profiler, google-python-cloud-debugger
google-auth==1.19.2        # via google-api-core, google-api-python-client, google-auth-httplib2, google-cloud-profiler, google-python-cloud-debugger
google-cloud-core==0.29.1  # via google-cloud-trace
google-cloud-profiler==1.0.8  # via -r requirements.in
google-cloud-trace==0.20.2  # via opencensus
#google-python-cloud-debugger==2.9  # via -r requirements.in
googleapis-common-protos==1.52.0  # via google-api-core
grpcio-health-checking==1.30.0  # via -r requirements.in
grpcio==1.30.0            # via -r requirements.in, google-api-core, grpcio-health-checking
httplib2==0.18.1          # via google-api-python-client, google-auth-httplib2
idna==2.10                 # via requests
opencensus[stackdriver]==0.1.10  # via -r requirements.in
protobuf==3.12.2          # via google-api-core, google-cloud-profiler, googleapis-common-protos, grpcio-health-checking
pyasn1-modules==0.2.4     # via google-auth
pyasn1==0.4.5             # via pyasn1-modules, rsa
python-json-logger==0.1.9  # via -r requirements.in
pytz==2018.9              # via google-api-core
pyyaml==5.1               # via google-python-cloud-debugger
requests==2.24.0          # via -r requirements.in, google-api-core
rsa==4.0                  # via google-auth
six==1.12.0               # via google-api-core, google-api-python-client, google-auth, google-auth-httplib2, google-python-cloud-debugger, grpcio, protobuf
uritemplate==3.0.0        # via google-api-python-client
urllib3==1.24.2           # via requests
#added requirements
setuptools
wheel
spacy==2.3.5
matplotlib
psutil
grpcio-tools
prometheus_client
